Expand API with a DB that can execute 'Scripts' and 'Queries'. ⤵️
This DB leverages prepared statements when configured thresholds are
encountered.

The DB type is the approach I'd like to use for new code.
Existing mappers can be more easily transitioned to the package-level
functions that were implemented in previous commits on this v3 branch.

Perhaps those functions will be moved to a separate package...